Refined WWWForm code.
Refined the code used to add fields, and send data to the online form. The new script "WebForm.cs" makes it easier to send data to a form.

AddField - works in the same way as before, but now can be accessed from outside the script. New fields can be added on the go by any other script that has access to this one.

ClearForm - removes all the field from the form.

SendForm - loops through all the fields in the dictionary, adds them to the form and sends them to the web server. The web address needs to be passed as a variable.

DisplayFormFields - loops through all the fields in the dictionary, and display the key + value.